scripts_and_support_files/requirements.txt
regex
google-cloud-storage
google-cloud-logging